Create separate eks and k8sClient files
