+++ pkg/kfapp/aws/identity.go
@@ -20,8 +20,13 @@ import (
 	"strings"
 )
 
-const AWS_DEFAULT_AUDIENCE = "sts.amazonaws.com"
-const AWS_TRUST_IDENTITY_SUBJECT = "system:serviceaccount:%s:%s"
+const (
+	AWS_DEFAULT_AUDIENCE               = "sts.amazonaws.com"
+	AWS_TRUST_IDENTITY_SUBJECT         = "system:serviceaccount:%s:%s"
+	AWS_SERVICE_ACCOUNT_ANNOTATION_KEY = "eks.amazonaws.com/role-arn"
+	AWS_IAM_ROLE_TAG_KEY               = "kubeflow/cluster-name"
+	AWS_IAM_ROLE_ARN                   = "arn:aws:iam::%s:role/%s"
+)
 
 // checkIdentityProviderExists will return true when the iam identity provider exists, it may return errors
 // if it was unable to call IAM API
@@ -43,40 +48,12 @@ func (aws *Aws) checkIdentityProviderExists(accountId, issuerURLWithoutProtocol 
 	return true, awssdk.StringValue(input.OpenIDConnectProviderArn), nil
 }
 
-// getIssueCAThumbprint will generate CAThumbprint from a given issuerURL
-func (aws *Aws) getIssueCAThumbprint(issuerURL string) (string, error) {
-	var issuerCAThumbprint string
-
-	client := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: true,
-			},
-			Proxy: http.ProxyFromEnvironment,
-		},
-	}
-
-	response, err := client.Get(issuerURL)
-	if err != nil {
-		return "", errors.Wrap(err, "connecting to issuer OIDC")
-	}
-
-	if response.TLS != nil {
-		if numCerts := len(response.TLS.PeerCertificates); numCerts >= 1 {
-			root := response.TLS.PeerCertificates[numCerts-1]
-			issuerCAThumbprint = fmt.Sprintf("%x", sha1.Sum(root.Raw))
-			return issuerCAThumbprint, nil
-		}
-	}
-	return "", errors.Errorf("unable to get OIDC issuer's certificate")
-}
-
 // createIdentityProvider create an OpenIDConnectProvider, it's one to one mapping to EKS cluster.
 func (aws *Aws) createIdentityProvider(issuerUrl string) (string, error) {
-	issuerCAThumbprint, err := aws.getIssueCAThumbprint(issuerUrl)
+	issuerCAThumbprint, err := getIssueCAThumbprint(issuerUrl)
 
 	oidcProviderInput := &iam.CreateOpenIDConnectProviderInput{
-		ClientIDList:   []*string{awssdk.String(DEFAULT_AUDIENCE)},
+		ClientIDList:   []*string{awssdk.String(AWS_DEFAULT_AUDIENCE)},
 		ThumbprintList: []*string{awssdk.String(issuerCAThumbprint)},
 		Url:            awssdk.String(issuerUrl),
 	}
@@ -90,51 +67,61 @@ func (aws *Aws) createIdentityProvider(issuerUrl string) (string, error) {
 	return awssdk.StringValue(output.OpenIDConnectProviderArn), nil
 }
 
-// DeleteIdentityProvider will delete the identity provider using IAM API, it may return an error the API call fails
+// DeleteIdentityProvider will delete the identity provider, it may return an error the API call fails
 func (aws *Aws) DeleteIdentityProvider(providerArn string) error {
 	input := &iam.DeleteOpenIDConnectProviderInput{
 		OpenIDConnectProviderArn: awssdk.String(providerArn),
 	}
 	if _, err := aws.iamClient.DeleteOpenIDConnectProvider(input); err != nil {
-		return errors.Wrap(err, "deleting OIDC provider")
+		return errors.Wrap(err, "deleting oidc provider")
 	}
 	return nil
 }
 
-func (aws *Aws) checkWebIdentityRoleExist(roleName string) error {
+// createOrUpdateWebIdentityRole creates an IAM role with trusted entity Web Identity if role doesn't exist
+func (aws *Aws) createOrUpdateWebIdentityRole(oidcProviderArn, issuerUrl, roleName, serviceAccountNamespace, serviceAccountName string) error {
 	input := &iam.GetRoleInput{
 		RoleName: awssdk.String(roleName),
 	}
 
+	// Don't need to update role, return immediately.
 	if _, err := aws.iamClient.GetRole(input); err != nil {
-		return err
+		// check non exist or other failures.
+		if aerr, ok := err.(awserr.Error); ok {
+			switch aerr.Code() {
+			case iam.ErrCodeNoSuchEntityException:
+				log.Infof("Role %v doesn't, creating it for KSA %s/%s", roleName, serviceAccountNamespace, serviceAccountName)
+			default:
+				return err
+			}
+		} else {
+			return err
+		}
+	} else {
+		log.Warnf("Role %v exists, skip creating role", roleName)
+		return nil
 	}
 
-	return nil
-}
-
-// createWebIdentityRole creates an IAM role with trusted entity Web Identity
-func (aws *Aws) createWebIdentityRole(oidcProviderArn, issuerUrlWithProtocol, roleName, namespace, ksa string) error {
+	// Create role
 	statement := MakeAssumeRoleWithWebIdentityPolicyDocument(oidcProviderArn, MapOfInterfaces{
 		"StringEquals": map[string][]string{
-			issuerUrlWithProtocol + ":aud": []string{AWS_DEFAULT_AUDIENCE},
-			issuerUrlWithProtocol + ":sub": []string{fmt.Sprintf(AWS_TRUST_IDENTITY_SUBJECT, namespace, ksa)},
+			issuerUrl + ":aud": []string{AWS_DEFAULT_AUDIENCE},
+			issuerUrl + ":sub": []string{fmt.Sprintf(AWS_TRUST_IDENTITY_SUBJECT, serviceAccountNamespace, serviceAccountName)},
 		},
 	})
 
 	assumeRolePolicyDocument := MakePolicyDocument(statement)
-
-	docInBytes, err := json.Marshal(assumeRolePolicyDocument)
+	document, err := json.Marshal(assumeRolePolicyDocument)
 	if err != nil {
-		return errors.Errorf("%v can not be marshal to bytes", docInBytes)
+		return errors.Errorf("%v can not be marshal to bytes", document)
 	}
 
 	roleInput := &iam.CreateRoleInput{
 		RoleName:                 awssdk.String(roleName),
-		AssumeRolePolicyDocument: awssdk.String(string(docInBytes)),
+		AssumeRolePolicyDocument: awssdk.String(string(document)),
 		Tags: []*iam.Tag{
 			{
-				Key: awssdk.String("kubeflow/cluster-name"),
+				Key: awssdk.String(AWS_IAM_ROLE_TAG_KEY),
 				// roleName is like kf-admin-clusterName
 				Value: awssdk.String(roleName[strings.LastIndex("roleName", "-")+1:]),
 			},
@@ -148,35 +135,17 @@ func (aws *Aws) createWebIdentityRole(oidcProviderArn, issuerUrlWithProtocol, ro
 	return nil
 }
 
-// deleteIAMRole delete an IAM role
-func (aws *Aws) deleteIAMRole(roleName string) error {
-	input := &iam.DeleteRoleInput{
-		RoleName: awssdk.String(roleName),
-	}
-
-	if _, err := aws.iamClient.DeleteRole(input); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func getIAMRoleNameFromIAMRoleArn(arn string) string {
-	return arn[strings.LastIndex(arn, "/")+1:]
-}
-
 // createOrUpdateK8sServiceAccount creates or updates k8s service account with annotation
-func (aws *Aws) createOrUpdateK8sServiceAccount(k8sClientset *clientset.Clientset, namespace, saName, accountId, iamRoleName string) error {
-	iamRoleArn := fmt.Sprintf("arn:aws:iam::%s:role/%s", accountId, iamRoleName)
-	existingSA, err := k8sClientset.CoreV1().ServiceAccounts(namespace).Get(saName, metav1.GetOptions{})
+func (aws *Aws) createOrUpdateK8sServiceAccount(k8sClientset *clientset.Clientset, serviceAccountNamespace, serviceAccountName, iamRoleArn string) error {
+	existingSA, err := k8sClientset.CoreV1().ServiceAccounts(serviceAccountNamespace).Get(serviceAccountName, metav1.GetOptions{})
 	if err == nil {
-		log.Infof("Service account %v already exists", saName)
+		log.Infof("Service account %v already exists", serviceAccountName)
 		if existingSA.Annotations == nil {
 			existingSA.Annotations = map[string]string{}
 		}
 
-		existingSA.Annotations["eks.amazonaws.com/role-arn"] = iamRoleArn
-		_, err = k8sClientset.CoreV1().ServiceAccounts(namespace).Update(existingSA)
+		existingSA.Annotations[AWS_SERVICE_ACCOUNT_ANNOTATION_KEY] = iamRoleArn
+		_, err = k8sClientset.CoreV1().ServiceAccounts(serviceAccountNamespace).Update(existingSA)
 		if err != nil {
 			return &kfapis.KfError{
 				Code:    int(kfapis.INTERNAL_ERROR),
@@ -186,14 +155,14 @@ func (aws *Aws) createOrUpdateK8sServiceAccount(k8sClientset *clientset.Clientse
 		return nil
 	}
 
-	log.Infof("Can not find service account, Creating service account %s/%s", namespace, saName)
-	_, err = k8sClientset.CoreV1().ServiceAccounts(namespace).Create(
+	log.Infof("Can not find existing service account, creating service account %s/%s", serviceAccountNamespace, serviceAccountName)
+	_, err = k8sClientset.CoreV1().ServiceAccounts(serviceAccountNamespace).Create(
 		&v1.ServiceAccount{
 			ObjectMeta: metav1.ObjectMeta{
-				Name:      saName,
-				Namespace: namespace,
+				Name:      serviceAccountName,
+				Namespace: serviceAccountNamespace,
 				Annotations: map[string]string{
-					"eks.amazonaws.com/role-arn": iamRoleArn,
+					AWS_SERVICE_ACCOUNT_ANNOTATION_KEY: iamRoleArn,
 				},
 			},
 		},
@@ -208,8 +177,8 @@ func (aws *Aws) createOrUpdateK8sServiceAccount(k8sClientset *clientset.Clientse
 	}
 }
 
-// updateRoleTrustIdentity add namespace/service to IAM Role trust entity
-func (aws *Aws) updateRoleTrustIdentity(serviceAccountNamespace string, serviceAccountName string, roleName string) error {
+// updateRoleTrustIdentity add namespace/serviceAccount to IAM Role trust entity
+func (aws *Aws) updateRoleTrustIdentity(roleName, serviceAccountNamespace, serviceAccountName string) error {
 	roleInput := &iam.GetRoleInput{
 		RoleName: awssdk.String(roleName),
 	}
@@ -219,7 +188,7 @@ func (aws *Aws) updateRoleTrustIdentity(serviceAccountNamespace string, serviceA
 		return err
 	}
 
-	// Seems AssumeRolePolicyDocument is URL encoded
+	// Seems AssumeRolePolicyDocument is URL encoded, decode string to get string policy document
 	decodeValue, err := url.QueryUnescape(awssdk.StringValue(output.Role.AssumeRolePolicyDocument))
 	if err != nil {
 		return err
@@ -241,6 +210,7 @@ func (aws *Aws) updateRoleTrustIdentity(serviceAccountNamespace string, serviceA
 	return nil
 }
 
+// getUpdatedAssumeRolePolicy creates a new policy document with a new ns/sa record
 func getUpdatedAssumeRolePolicy(policyDocument, serviceAccountNamespace, serviceAccountName string) (string, error) {
 	var oldDoc MapOfInterfaces
 	json.Unmarshal([]byte(policyDocument), &oldDoc)
@@ -289,12 +259,58 @@ func getUpdatedAssumeRolePolicy(policyDocument, serviceAccountNamespace, service
 	return string(newPolicyDoc), nil
 }
 
+// getIssueCAThumbprint will generate CAThumbprint from a given issuerURL, this is used to create an oidc provider
+func getIssueCAThumbprint(issuerURL string) (string, error) {
+	var issuerCAThumbprint string
+
+	client := &http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{
+				InsecureSkipVerify: true,
+			},
+			Proxy: http.ProxyFromEnvironment,
+		},
+	}
+
+	response, err := client.Get(issuerURL)
+	if err != nil {
+		return "", errors.Wrap(err, "connecting to issuer OIDC")
+	}
+
+	if response.TLS != nil {
+		if numCerts := len(response.TLS.PeerCertificates); numCerts >= 1 {
+			root := response.TLS.PeerCertificates[numCerts-1]
+			issuerCAThumbprint = fmt.Sprintf("%x", sha1.Sum(root.Raw))
+			return issuerCAThumbprint, nil
+		}
+	}
+	return "", errors.Errorf("unable to get OIDC issuer's certificate")
+}
+
+// deleteIAMRole delete an IAM role
+func (aws *Aws) deleteIAMRole(roleName string) error {
+	input := &iam.DeleteRoleInput{
+		RoleName: awssdk.String(roleName),
+	}
+
+	if _, err := aws.iamClient.DeleteRole(input); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// getIAMRoleNameFromIAMRoleArn converts roleArn to roleName
+func getIAMRoleNameFromIAMRoleArn(arn string) string {
+	return arn[strings.LastIndex(arn, "/")+1:]
+}
+
 // getIssuerUrlFromRoleArn parse issuerUrl from Arn: arn:aws:iam::${accountId}:oidc-provider/${issuerUrl}
 func getIssuerUrlFromRoleArn(arn string) string {
 	return arn[strings.Index(arn, "/")+1:]
 }
 
-// MakeAssumeRoleWithWebIdentityPolicyDocument constructs a trust policy for given a web identity provider with given conditions
+// MakeAssumeRoleWithWebIdentityPolicyDocument constructs a trust policy statement for given web identity provider with given conditions
 func MakeAssumeRoleWithWebIdentityPolicyDocument(providerARN string, condition MapOfInterfaces) MapOfInterfaces {
 	return MapOfInterfaces{
 		"Effect": "Allow",
@@ -306,7 +322,7 @@ func MakeAssumeRoleWithWebIdentityPolicyDocument(providerARN string, condition M
 	}
 }
 
-// MakePolicyDocument constructs a policy with given statements
+// MakePolicyDocument constructs a policy document with given statements
 func MakePolicyDocument(statements ...MapOfInterfaces) MapOfInterfaces {
 	return MapOfInterfaces{
 		"Version":   "2012-10-17",
