+++ pkg/utils/awsutil.go
@@ -17,8 +17,10 @@ limitations under the License.
 package utils
 
 import (
+	"fmt"
 	awssdk "github.com/aws/aws-sdk-go/aws"
 	"os/exec"
+	"regexp"
 
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sts"
@@ -64,7 +66,7 @@ func CheckCommandExist(commandName string) error {
 	return nil
 }
 
-// GetEksctlVersion return eksctl version on user's machine
+// GetEksctlVersion return eksctl version on user's environment
 func GetEksctlVersion() error {
 	log.Infof("Running `eksctl version` ...")
 	output, err := exec.Command("eksctl", "version").Output()
@@ -74,8 +76,14 @@ func GetEksctlVersion() error {
 		return err
 	}
 
-	log.Infof("output: %v", string(output))
 	// [ℹ]  version.Info{BuiltAt:"", GitCommit:"", GitTag:"0.1.32"}
-	// We'd like to extract 0.1.32 and compare with minimum version we support.
+	r := regexp.MustCompile("[0-9]+.[0-9]+.[0-9]+")
+	matchGroups := r.FindStringSubmatch(string(output))
+
+	if len(matchGroups) == 0 {
+		return fmt.Errorf("can not find eksctl version from %v", string(output))
+	}
+
+	log.Infof("eksctl version: %s", matchGroups[0])
 	return nil
 }
